marketplace/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
CacheControl==0.14.4
cachetools==6.2.0
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.3
distlib==0.3.9
dj-config-url==0.1.1
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.18.0
firebase_admin==7.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.8.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
msgpack==1.1.2
mysql-connector==2.2.9
mysql-connector-python==9.3.0
packaging==25.0
pillow==11.3.0
pipenv==2025.0.3
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.1
pytest-mock==3.14.1
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.31.2
whitenoise==6.11.0
